+++ LHCbDIRAC/DataManagementSystem/DB/StorageUsageDB.py
@@ -525,7 +525,7 @@ class StorageUsageDB(DB):
   def purgeOutdatedEntries(self, rootDir=False, outdatedSeconds=86400, preserveDirsList=None):
     preserveDirsList = preserveDirsList if preserveDirsList else []
     try:
-      outdatedSeconds = max(1, long(outdatedSeconds))
+      outdatedSeconds = max(1, int(outdatedSeconds))
     except ValueError:
       return S_ERROR("Ooutdated seconds needs to be a number")
 
@@ -582,7 +582,7 @@ class StorageUsageDB(DB):
       sqlCond.append("d.Path LIKE '%%/%s%%'" % fileType)
     if production:
       try:
-        sqlCond.append("d.Path LIKE '%%/%08.d%%'" % long(production))
+        sqlCond.append("d.Path LIKE '%%/%08.d%%'" % int(production))
       except ValueError:
         return S_ERROR("production has to be a number")
     if SEs:
@@ -604,7 +604,7 @@ class StorageUsageDB(DB):
       return result
     usageDict = {}
     for gf, size, files in result['Value']:
-      usageDict[gf] = {'Size': long(size), 'Files': long(files)}
+      usageDict[gf] = {'Size': int(size), 'Files': int(files)}
     return S_OK(usageDict)
 
   def getStorageSummary(self, path, fileType=False, production=False, SEs=[]):
@@ -640,7 +640,7 @@ class StorageUsageDB(DB):
       return result
     userStorage = {}
     for row in result['Value']:
-      userStorage[row[0]] = long(row[1])
+      userStorage[row[0]] = int(row[1])
     return S_OK(userStorage)
 
   def getUserSummaryPerSE(self, userName=False):
@@ -662,7 +662,7 @@ class StorageUsageDB(DB):
       seName = row[1]
       if userName not in userData:
         userData[userName] = {}
-      userData[userName][seName] = {'Size': long(row[2]), 'Files': long(row[3])}
+      userData[userName][seName] = {'Size': int(row[2]), 'Files': int(row[3])}
     return S_OK(userData)
 
   def getDirectorySummaryPerSE(self, directory):
@@ -679,15 +679,15 @@ class StorageUsageDB(DB):
       seName = row[0]
       if seName not in data:
         data[seName] = {}
-      data[seName] = {'Size': long(row[1]), 'Files': long(row[2])}
+      data[seName] = {'Size': int(row[1]), 'Files': int(row[2])}
     return S_OK(data)
 
   def publishTose_STSummary(self, site, spaceToken, totalSize, totalFiles, storageDumpLastUpdate):
     """Publish total size and total files extracted from the storage dumps to
     the se_STSummary."""
     try:
-      sqlTotalSize = long(totalSize)
-      sqlTotalFiles = long(totalFiles)
+      sqlTotalSize = int(totalSize)
+      sqlTotalFiles = int(totalFiles)
     except ValueError as e:
       return S_ERROR("Values must be ints: %s" % repr(e))
     sqlSpaceToken = self._escapeString(spaceToken)['Value']
@@ -799,7 +799,7 @@ class StorageUsageDB(DB):
         data[thisRun] = {}
         for row in result['Value']:
           seName = row[0]
-          data[thisRun][seName] = {'Size': long(row[1]), 'Files': long(row[2])}
+          data[thisRun][seName] = {'Size': int(row[1]), 'Files': int(row[2])}
     else:
       sqlCmd = "SELECT su.SEName, SUM(su.Size), SUM(su.Files)  FROM su_Directory AS d, su_SEUsage AS su WHERE " \
           "d.DID=su.DID and d.Path LIKE '/lhcb/data/%%/RAW/%%/%%/%%/%d/' GROUP BY su.SEName" % (run)
@@ -809,7 +809,7 @@ class StorageUsageDB(DB):
         return S_ERROR(result)
       for row in result['Value']:
         seName = row[0]
-        data[seName] = {'Size': long(row[1]), 'Files': long(row[2])}
+        data[seName] = {'Size': int(row[1]), 'Files': int(row[2])}
 
     return S_OK(data)
 
@@ -865,7 +865,7 @@ class StorageUsageDB(DB):
       sqlCond.append("Path LIKE '%%/%s%%'" % fileType)
     if production:
       try:
-        sqlCond.append("Path LIKE '%%/%08.d%%'" % long(production))
+        sqlCond.append("Path LIKE '%%/%08.d%%'" % int(production))
       except ValueError:
         return S_ERROR("production has to be a number")
     return sqlCond
@@ -882,7 +882,7 @@ class StorageUsageDB(DB):
       return result
     usageDict = {}
     for gf, size, files in result['Value']:
-      usageDict[gf] = {'Size': long(size), 'Files': long(files)}
+      usageDict[gf] = {'Size': int(size), 'Files': int(files)}
     return S_OK(usageDict)
 
   def getSummary(self, path, fileType=False, production=False):
