+++ LHCbDIRAC/ProductionManagementSystem/DB/ProductionRequestDB.py
@@ -167,7 +167,7 @@ class ProductionRequestDB(DB):
         rec[x] = requestDict[x]  # set only known not empty fields
     if rec['NumberOfEvents']:  # Set RealNumberOfEvents if specified
       try:
-        num = long(rec['NumberOfEvents'])
+        num = int(rec['NumberOfEvents'])
         if num > 0:
           rec['RealNumberOfEvents'] = num
       except ValueError:
@@ -213,7 +213,7 @@ class ProductionRequestDB(DB):
 
     if rec['MasterID']:  # have to check ParentID and MasterID consistency
       try:
-        masterID = long(rec['MasterID'])
+        masterID = int(rec['MasterID'])
       except ValueError:
         self.lock.release()
         return S_ERROR('MasterID is not a number')
@@ -221,7 +221,7 @@ class ProductionRequestDB(DB):
         self.lock.release()
         return S_ERROR('MasterID can not be without ParentID')
       try:
-        parentID = long(rec['ParentID'])
+        parentID = int(rec['ParentID'])
       except ValueError:
         self.lock.release()
         return S_ERROR('ParentID is not a number')
@@ -240,7 +240,7 @@ class ProductionRequestDB(DB):
         return S_ERROR("Only request author can add subrequests")
     elif rec['ParentID']:
       try:
-        parentID = long(rec['ParentID'])
+        parentID = int(rec['ParentID'])
       except ValueError:
         self.lock.release()
         return S_ERROR('ParentID is not a number')
@@ -347,7 +347,7 @@ class ProductionRequestDB(DB):
       filterIn = {}
     try:  # test parameters
       for x in requestIDList:
-        y = long(x)
+        y = int(x)
     except ValueError:
       return S_ERROR("Bad parameters (all request IDs must be numbers)")
     idFilter = False
@@ -738,7 +738,7 @@ class ProductionRequestDB(DB):
     if 'NumberOfEvents' in update:  # Update RealNumberOfEvents if specified
       num = 0
       try:
-        num = long(rec['NumberOfEvents'])
+        num = int(rec['NumberOfEvents'])
         if num < 0:
           num = 0
       except ValueError:
@@ -828,7 +828,7 @@ class ProductionRequestDB(DB):
     Available is New and Rejected states only
     """
     try:
-      requestID = long(requestID)
+      requestID = int(requestID)
     except ValueError:
       return S_ERROR('RequestID is not a number')
     self.lock.acquire()  # transaction begin ?? may be after connection ??
@@ -977,7 +977,7 @@ class ProductionRequestDB(DB):
     # Clear RealNumberOfEvents if required
     try:
       num = 0
-      num = long(rec['NumberOfEvents'])
+      num = int(rec['NumberOfEvents'])
       if num < 0:
         num = 0
     except ValueError:
@@ -1034,7 +1034,7 @@ class ProductionRequestDB(DB):
       if not result['OK']:
         return result
 
-    return S_OK(long(newRequestID))
+    return S_OK(int(newRequestID))
 
   def duplicateProductionRequest(self, requestID, creds, clearpp):
     """Duplicate production request with all it's subrequests (but without
@@ -1045,7 +1045,7 @@ class ProductionRequestDB(DB):
     (of the master) are cleaned.
     """
     try:
-      requestID = long(requestID)
+      requestID = int(requestID)
     except ValueError:
       return S_ERROR('RequestID is not a number')
     self.lock.acquire()  # transaction begin ?? may be after connection ??
@@ -1123,14 +1123,14 @@ class ProductionRequestDB(DB):
     state can request the split.
     """
     try:
-      requestID = long(requestID)
+      requestID = int(requestID)
     except ValueError:
       return S_ERROR('RequestID is not a number')
     if not splitlist:
       return S_ERROR('Split list is empty')
     isplitlist = []
     try:
-      isplitlist = [long(x) for x in splitlist]
+      isplitlist = [int(x) for x in splitlist]
     except ValueError:
       return S_ERROR('RequestID in split list is not a number')
 
@@ -1166,9 +1166,9 @@ class ProductionRequestDB(DB):
     keeplist = []
     for ch in result['Value']:
       if ch[0] in isplitlist and ch[1] == requestID:
-        fisplitlist.append(long(ch[0]))
+        fisplitlist.append(int(ch[0]))
       elif ch[1]:
-        keeplist.append(long(ch[0]))
+        keeplist.append(int(ch[0]))
     if len(isplitlist) != len(fisplitlist):
       self.lock.release()
       return S_ERROR('Requested for spliting subrequests are no longer exist')
@@ -1194,7 +1194,7 @@ class ProductionRequestDB(DB):
     if not result['OK']:
       self.lock.release()
       return result
-    newRequestID = long(result['Value'][0][0])
+    newRequestID = int(result['Value'][0][0])
 
     # Move subrequests (!! Errors are not fatal !!)
     rsplitlist = []
@@ -1252,7 +1252,7 @@ class ProductionRequestDB(DB):
     """
     try:
       for x in self.progressFields:
-        pdict[x] = long(pdict[x])
+        pdict[x] = int(pdict[x])
     except ValueError:
       return S_ERROR('Bad parameters')
 
@@ -1334,8 +1334,8 @@ class ProductionRequestDB(DB):
     # check parameters
     try:
       for x in update:
-        x['ProductionID'] = long(x['ProductionID'])
-        x['BkEvents'] = long(x['BkEvents'])
+        x['ProductionID'] = int(x['ProductionID'])
+        x['BkEvents'] = int(x['BkEvents'])
     except ValueError:
       return S_ERROR('Bad parameters')
     except TypeError:
@@ -1373,7 +1373,7 @@ class ProductionRequestDB(DB):
         continue
       del res['SimCondDetail']
       try:
-        num = long(res['RealNumberOfEvents'])
+        num = int(res['RealNumberOfEvents'])
       except ValueError:
         num = 0
       except TypeError:
@@ -1388,8 +1388,8 @@ class ProductionRequestDB(DB):
     # check parameters
     try:
       for x in update:
-        x['RequestID'] = long(x['RequestID'])
-        x['RealNumberOfEvents'] = long(x['RealNumberOfEvents'])
+        x['RequestID'] = int(x['RequestID'])
+        x['RealNumberOfEvents'] = int(x['RealNumberOfEvents'])
     except ValueError:
       return S_ERROR('Bad parameters')
     except TypeError:
